remove get next line, rework socket logic
