+++ inc/socket.h
@@ -1,23 +1,18 @@
 #ifndef SOCKET_H
 # define SOCKET_H
 
-# include <sys/socket.h>
-# include <sys/select.h>
-# include <netinet/in.h>
-# include <stdlib.h>
-# include <stdio.h>
-# include <unistd.h>
-# include <string.h>
-# include <arpa/inet.h>
-# include <fcntl.h>
-# include <errno.h>
-# include <stdbool.h>
-
-# ifndef BUFFER_SIZE
-#  define BUFFER_SIZE 2048
-# endif
-
-char	*get_next_line(int fd);
+#include <sys/socket.h>
+#include <sys/select.h>
+#include <netinet/in.h>
+#include <stdbool.h>
+#include <netdb.h>
+#include <arpa/inet.h>
+#include <fcntl.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <errno.h>
+#include <unistd.h>
 
 int					ft_init_socket(struct sockaddr_in addr);
 int					ft_send_socket(const int socket_fd, const char *msg);
